Enforce that if a stats_server_uri is provided, we can connect to it
