+++ metta/setup/tools/local/kind.py
@@ -4,9 +4,9 @@ from pathlib import Path
 from typing import Callable
 
 from devops.docker.push_image import push_image
+from metta.app_backend.clients.base_client import get_machine_token
 from metta.common.util.constants import DEV_STATS_SERVER_URI, METTA_AWS_ACCOUNT_ID, METTA_AWS_REGION
 from metta.common.util.fs import get_repo_root
-from metta.common.util.stats_client_cfg import get_machine_token
 from metta.setup.utils import error, info, success
 
 repo_root = get_repo_root()
